+++ NEWS.md
@@ -6,6 +6,8 @@
 
 ## Enhancements
 * `biplot()` allows to produce a symetric CA biplot.
+* Set graphical parameters as arguments in `biplot()`.
+* Add automatic legend in `biplot()`.
 
 ## Internals
 * Import **arkhe**.
@@ -13,7 +15,6 @@
 # dimensio 0.4.1
 ## Bugfixes & changes
 * Put `...` after required arguments.
-* Suggest **ggplot2** instead of import.
 
 ## Enhancements
 * `viz_individuals()`, `viz_row()`, `viz_variables()`, `viz_columns()` and `biplot()` gained new arguments allowing set the x and y limits of the plot.